Added an ignore list of packages
Had an issue with yay and yay-debug getting removed since we installed
it without adding them to the packages.conf file.

Even with adding them to the AUR section of the conf file, they still were
getting removed.

When adding them to the packages section, they were erroring out saying
that the package did not exist in pacman.